+++ packages/middleware-express-v4/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @armory-sh/middleware-express-v4
 
+## 0.3.15
+
+### Patch Changes
+
+- v2 Only, Cleanup
+- Updated dependencies
+  - @armory-sh/base@0.2.14
+
 ## 0.3.14
 
 ### Patch Changes