+++ packages/middleware-express/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @armory-sh/middleware-express
 
+## 0.4.2
+
+### Patch Changes
+
+- v2 Only, Cleanup
+- Updated dependencies
+  - @armory-sh/base@0.2.14
+
 ## 0.4.1
 
 ### Patch Changes